% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.text.R
\name{bartext}
\alias{bartext}
\title{Character-based Barplot}
\usage{
bartext(x, num.lines = 1, ...)
}
\arguments{
\item{x}{A vector containing values for the bars}

\item{num.lines}{The number of lines to use. Single-line plots will be rendered with different characters within the single line.}
}
\description{
\code{bartext} replaces the barplot function for character environments.
}

