% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_biterms.R
\docType{data}
\name{example_btm}
\alias{example_btm}
\title{Example Biterm Topic Model}
\description{
The object is a BTM topic model created with the BTM package.
It was created on a subset of all CRAN packages, namely package which
are part of the NaturalLanguageProcessing and MachineLearning task views.\cr
Timepoint of creation was 2020-04-10.
}
\examples{
\dontshow{if(require(BTM))
\{
}
library(BTM)
data(example_btm, package = 'textplot')
example_btm
str(example_btm)
\dontshow{
\}
# End of main if statement running only if the required packages are installed
}
}
