% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_embedding_2d.R
\docType{data}
\name{example_embedding}
\alias{example_embedding}
\title{Example word embedding matrix}
\description{
A matrix with 25-dimensional word embeddings, constructed upon the be_parliament_2020 dataset in the doc2vec R package
}
\examples{
data(example_embedding, package = 'textplot')
head(example_embedding)
}
