% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_embedding_2d.R
\docType{data}
\name{example_embedding_clusters}
\alias{example_embedding_clusters}
\title{Example words emitted in a ETM text clustering model}
\description{
Example words emitted in a ETM text clustering model constructed upon the be_parliament_2020 dataset in the doc2vec R package
}
\examples{
data(example_embedding_clusters, package = 'textplot')
head(example_embedding_clusters)
terminology <- split(example_embedding_clusters, example_embedding_clusters$cluster)
lapply(terminology, head, n = 5)
}
