% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_dependencies.R
\docType{data}
\name{example_udpipe}
\alias{example_udpipe}
\title{Example annotation of text using udpipe}
\description{
The object is a data.frame of the annotation of the text:
"UDPipe provides tokenization, tagging, lemmatization and dependency parsing of raw text"
}
\examples{
data(example_udpipe)
str(example_udpipe)
}
