% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_dispersion.R
\name{lexical_dispersion}
\alias{lexical_dispersion}
\title{Lexical Dispersion}
\usage{
lexical_dispersion(text.var, term.list, grouping.var = NULL, rm.var = NULL,
  group.names, time.names, ignore.case = TRUE, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{term.list}{A vector of quoted terms or a named list of quoted terms.
If the latter terms will be combined into a single unified theme named
according to the list names.  Note that terms within the vectors of the list
cannot be duplicated.}

\item{grouping.var}{The grouping variables.  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{rm.var}{The repeated measures variables.  Default \code{NULL} generates
one facet for all text.  Also takes a single repeated measures variable or
a list of 1 or more grouping variables.}

\item{group.names}{A vector of names that corresponds to group.var.  Generally
for internal use.}

\item{time.names}{A vector of names that corresponds to rm.var.  Generally
for internal use.}

\item{ignore.case}{logical.  If \code{TRUE} matching will be done without
regard to case.}

\item{\ldots}{Ignored.}
}
\value{
Returns dispersion plot data (location of the terms) and default
plots the dispersion plot.
}
\description{
Generate a lexical dispersion dataset (location of terms).  Typically the
user will want to use \code{lexical_dispersion_plot} directly but this allows
the return of the data that generates the plot.
}
\note{
The match.terms is character sensitive.  Spacing is an important way
to grab specific words and requires careful thought.  Using "read" will find
the words "bread", "read" "reading", and "ready".  If you want to search
for just the word "read" you'd supply a vector of c(" read ", " reads",
" reading", " reader").
}
\examples{
x <- lexical_dispersion(sam_i_am, c('not', 'eat', 'sam'))
plot(x)
}
\keyword{dispersion}
