% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{hilighter_cols}
\alias{hilighter_cols}
\alias{view_cols}
\alias{assert_hex}
\title{Palettes and Palette Viewing}
\format{An object of class \code{character} of length 10.}
\usage{
hilighter_cols

view_cols(x)

assert_hex(x)
}
\arguments{
\item{x}{A vector of colors (either R \code{colors()} or hexadecimal).}
}
\value{
\code{view_cols} plots the colors as bars

\code{assert_hex} returns a vector of hexadecimal strings
}
\description{
\code{hilighter_cols} - A palette of 10 hilighter colors.

\code{view_cols} - Plots vector of colors.

\code{assert_hex} - Convert colors to hexadecimal.
}
\examples{
view_cols(hilighter_cols)
view_cols(c("#FFA500", "#B3B3B3", "#FFFF00", 'red', 'blue', 'black'))
}
\keyword{datasets}
