% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{plot.hilight}
\alias{plot.hilight}
\title{Plots a hilight Object}
\usage{
\method{plot}{hilight}(x, path = file.path(tempdir(), "hilight.html"),
  open = TRUE, ...)
}
\arguments{
\item{x}{A \code{hilight} object.}

\item{path}{A path ending in .html to output the plot to.}

\item{open}{logical  If \code{TRUE} and interactive, the plot will be opened
in the browser or RStudio.}

\item{\ldots}{ignored.}
}
\description{
Plots a hilight object
}
