% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_dispersion.R
\name{plot.lexical_dispersion}
\alias{plot.lexical_dispersion}
\title{Plots a lexical_dispersion object}
\usage{
\method{plot}{lexical_dispersion}(x, color = "blue", bg.color = "grey90",
  horiz.color = "grey85", total.color = "black", symbol = "|",
  title = "Lexical Dispersion Plot", rev.factor = TRUE, wrap = "'",
  xlab = "Dialogue (Words)", ylab = NULL, size = 4, scales = "free",
  space = "free", ...)
}
\arguments{
\item{x}{The lexical_dispersion object.}

\item{color}{The color of the word symbols.}

\item{bg.color}{The background color.}

\item{horiz.color}{The color of the horizontal tracking stripe.  Use
\code{horiz.color = bg.color} to eliminate.}

\item{total.color}{The color to use for summary `all` group.  If \code{NULL}
totals are dropped.}

\item{symbol}{The word symbol.  Default is \code{"|"}.}

\item{title}{Title of the plot}

\item{rev.factor}{logical.  If \code{TRUE} reverses the plot order of the
factors.}

\item{wrap}{a character to wrap around the words (enables the reader to
visualize spaces).  Default is \code{"'"}, use \code{""} to remove.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{size}{The size of the plotting symbol.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default), free
(\code{"free"}), or free in one dimension (\code{"free_x"}, \code{"free_y"})}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the y
scale; if \code{"free_x"} their width will be proportional to the length of
the x scale; or if \code{"free"} both height and width will vary.}

\item{\ldots}{Ignored.}
}
\description{
Plots a lexical_dispersion object.
}
