% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_tree.R
\name{plot.word_tree}
\alias{plot.word_tree}
\title{Plots a word_tree Object}
\usage{
\method{plot}{word_tree}(x, path = file.path(tempdir(), "word_tree.html"),
  width = 900, height = 600, open = TRUE, ...)
}
\arguments{
\item{x}{A \code{word_tree} object.}

\item{path}{A path ending in .html to output the plot to.}

\item{width}{The width of the plot in px for browsers.}

\item{height}{The height of the plot in px for browsers.}

\item{open}{logical  If \code{TRUE} and interactive, the plot will be opened 
in the browser or RStudio.}

\item{\ldots}{ignored.}
}
\description{
Plots a word_tree object
}
