% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterText.R
\name{scatterText}
\alias{scatterText}
\title{Character-based Scatterplot}
\usage{
scatterText(x, y, char.dim = 5, style = 1)
}
\arguments{
\item{x}{the x coordinates of points in the plot.}

\item{y}{the y coordinates of points in the plot,}

\item{char.dim}{size of the plot in characters. Either a single number that defines the width and height or a vector of size two describing x and y size}
}
\description{
\code{scatterText} produces a character-based scatter plot, in which the points
are represented by characters. This function supports various visual styles to
render the plot.
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
scatterText(x,y, char.dim=c(10,5))

}
\author{
Andreas M. Brandmaier
}

