% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{textplot-package}
\alias{textplot}
\alias{textplot-package}
\title{A package for creating character-based plots}
\description{
General plotting functions are found in \code{\link{scatterText}} and \code{\link{barText}}.
}

