% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot-package.R
\docType{package}
\name{textplot}
\alias{textplot}
\alias{package-textplot}
\alias{textplot-package}
\title{What the Package Does (one line)}
\description{
Analyze text data via serveral common text plotting methods.  Methods include
lexical dispersion plots, tree maps, speech networks, speech Gantt charts,
text hilighting, word clouds, etc.
}
