% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_corlines.R
\name{textplot_correlation_lines_attrs}
\alias{textplot_correlation_lines_attrs}
\title{Document/Term Correlation Plot graphical attributes}
\usage{
textplot_correlation_lines_attrs(fontsize = 25)
}
\arguments{
\item{fontsize}{size of the font. Defaults to 25}
}
\value{
a list with graph visualisation elements used by \code{\link{textplot_correlation_lines}}
}
\description{
Document/Term Correlation Plot graphical attributes
}
\examples{
textplot_correlation_lines_attrs()
}
