% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_entites.R
\name{textplot_entity.data.frame}
\alias{textplot_entity.data.frame}
\title{Plot output of a named-entity extraction}
\usage{
\method{textplot_entity}{data.frame}(
  x,
  doc_id = "doc_id",
  token = "token",
  entity_type = "entity_type"
)
}
\arguments{
\item{x}{a data.frame as returned by a call to \code{\link[udpipe]{udpipe}} containing 1 sentence}

\item{title}{character string with the title to use in the plot}

\item{subtitle}{character string with the title to use in the plot}

\item{vertex_color}{character with the color of the label of each node. Defaults to darkgreen.}

\item{edge_color}{character with the color of the edges between the nodes. Defaults to red.}

\item{size}{size of the labels in the plot. Defaults to 3.}

\item{base_family}{character passed on to \code{theme_void} setting the base font family}

\item{...}{not used yet}
}
\value{
an object of class ggplot
}
\description{
Plot output of a named-entity extraction.
This plot highlight the text with the associated entity,
the words, the parts of speech tag and the dependency relationship between the words.
}
\examples{
\dontshow{
if(require(spacyr) && require(ggplot2) && require(igraph))
\{
}
library(spqcyr)
library(ggplot2)
library(igraph)
\donttest{
txt1 <- c(doc1 = "I would have accepted without question the information that Gatsby sprang from the swamps of Louisiana or from the lower East Side of New York.",
          doc2 = "I graduated from New Haven in 1915, just a quarter of a century after my father, and a little later I participated in that delayed Teutonic migration known as the Great War.")

entity_df <- entity_consolidate(spacy_parse(txt1), concatenator = " ")
textplot_entity(entity_df)
}
}
\seealso{
\code{\link[spacyr]{spacyr}}
}
