% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_tree.R
\name{word_tree}
\alias{word_tree}
\title{Word Trees}
\usage{
word_tree(text.var, word, grouping.var = NULL, path = file.path(tempdir(),
  "treemap.html"), open = TRUE, width = 900, height = 600,
  ignore.case = TRUE, ...)
}
\arguments{
\item{text.var}{The text string variable.}

\item{word}{A character string with the word to search for.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.  If \code{TRUE} an \code{id} variable is
used with a \code{seq_along} the \code{text.var}.}

\item{path}{A path ending in .html to output the plot to.}

\item{open}{logical.  If \code{TRUE} the output .html file is opened via
\code{\link[utils]{browseURL}}.}

\item{width}{The width of the plot in px.}

\item{height}{The height of the plot in px.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern initial matching is
case sensitive and if \code{TRUE}, case is ignored during initial matching.}

\item{\ldots}{ignored.}
}
\value{
Makes a .html file and returns the path invisibly.
}
\description{
A port of Google Chart's word tree \url{https://developers.google.com/chart/interactive/docs/gallery/wordtree}.
}
\examples{
\dontrun{
word_tree(sam_i_am, word = 'I', open = TRUE)
word_tree(sam_i_am, word = 'do', open = TRUE)
word_tree(sam_i_am, word = 'not', open = TRUE)

with(presidential_debates_2012, word_tree(dialogue, word = 'I',
    list(person, time), open = TRUE))
with(presidential_debates_2012, word_tree(dialogue, word = 'america', l
    ist(person, time), open = TRUE, path = 'treemap.html'))
with(presidential_debates_2012, word_tree(dialogue, word = 'He',
    list(person, time), open = TRUE, path = 'treemap.html'))
with(presidential_debates_2012, word_tree(dialogue, word = 'we',
    list(person, time), open = TRUE, path = 'treemap.html', height = 1200))
with(presidential_debates_2012, word_tree(dialogue, word = 'our',
    list(person, time), open = TRUE, path = 'treemap.html', height = 1200))
}
}
\references{
\url{https://developers.google.com/chart/interactive/docs/gallery/wordtree}
}
\keyword{wordtree}
