% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_tree.R
\name{word_tree}
\alias{word_tree}
\title{Word Trees}
\usage{
word_tree(text.var, word, grouping.var = NULL, ignore.case = TRUE, ...)
}
\arguments{
\item{text.var}{The text string variable.}

\item{word}{A character string with the word to search for.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.  If \code{TRUE} an \code{id} variable is
used with a \code{seq_along} the \code{text.var}.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern initial matching is
case sensitive and if \code{TRUE}, case is ignored during initial matching.}

\item{\ldots}{ignored.}
}
\value{
Makes a .html file and returns the path invisibly.
}
\description{
A port of Google Chart's word tree \url{https://developers.google.com/chart/interactive/docs/gallery/wordtree}.
}
\examples{
\dontrun{
library(dplyr)

word_tree(sam_i_am, word = 'I') \%>\%
    plot()
word_tree(sam_i_am, word = 'do') \%>\%
    plot()
word_tree(sam_i_am, word = 'not') \%>\%
    plot()

with(presidential_debates_2012, word_tree(dialogue, word = 'I',
    list(person, time))) \%>\%
    plot()
with(presidential_debates_2012, word_tree(dialogue, word = 'america', 
    list(person, time), path = 'treemap.html')) \%>\%
    plot()
with(presidential_debates_2012, word_tree(dialogue, word = 'He',
    list(person, time), path = 'treemap.html')) \%>\%
    plot()
with(presidential_debates_2012, word_tree(dialogue, word = 'we',
    list(person, time), path = 'treemap.html')) \%>\%
    plot(height = 1200)
with(presidential_debates_2012, word_tree(dialogue, word = 'our',
    list(person, time), path = 'treemap.html')) \%>\%
    plot(height = 1200)
}
}
\references{
\url{https://developers.google.com/chart/interactive/docs/gallery/wordtree}
}
\keyword{wordtree}
